+++ tests/test_utils/mutate_helpers/test.py
@@ -11,7 +11,13 @@ R_FILE = Path(__file__).parent.parent.parent.parent.joinpath(
 )
 
 
-def run_fun(fun: str, fun_args: dict, df_values: dict, mutate: bool):
+def run_fun(
+    fun: str,
+    fun_args: dict,
+    df_values: dict,
+    mutate: bool,
+    if_else: bool = False,
+):
     df_r_repr = ", ".join(
         [f"{col} = {r(values)}" for col, values in df_values.items()]
     )
@@ -27,6 +33,18 @@ def run_fun(fun: str, fun_args: dict, df_values: dict, mutate: bool):
 
             df <- tibble({df_r_repr})
             out <- df %>% mutate(out = {fun}(., {fun_r_repr})) %>% pull(out)
+            cat(out)
+        """
+    elif if_else:
+        rcode = f"""
+            source("{R_FILE}")
+
+            df <- tibble({df_r_repr})
+            out <- df %>%
+                mutate(
+                    out = if_else(CDR3.aa %in% {fun}(., {fun_r_repr}), "X", "Y")
+                ) %>%
+                pull(out)
             cat(out)
         """
     else:
@@ -47,10 +65,11 @@ def run(
     expected: str,
     fun: str,
     mutate: bool = False,
+    if_else: bool = False,
     **fun_args: Any,
 ):
     print(f">>> TESTING {fun} - {fun_args}")
-    out = run_fun(fun, fun_args, df_values, mutate=mutate)
+    out = run_fun(fun, fun_args, df_values, mutate=mutate, if_else=if_else)
     assert out == expected, f"{out} != {expected}"
     print("    PASSED")
     print("")
@@ -143,4 +162,13 @@ if __name__ == "__main__":
         fun="vanished",
         group_by='"Source"',
         idents='c("Tumor", "Normal")',
+    )
+    run(
+        df_values=df_values,
+        expected="Y X X Y Y Y Y X X X Y Y",
+        fun="vanished",
+        group_by="Source",
+        idents='c("Tumor", "Normal")',
+        uniq='FALSE',
+        if_else=True,
     )